o Portability: size_t might not be formattable with %d.   Cast to a known long enough type for which there is a   formatting specifier.
git-svn-id: svn://repo.hu/pcb-rnd/trunk@3151 1d61de99-560f-4dba-8957-1e505d42358b

+++ src_plugins/hid_lesstif/main.c
@@ -2687,7 +2687,7 @@ static Boolean idle_proc(XtPointer dummy)
 
 		if (old_nrats != ratlist_length(&PCB->Data->Rat)) {
 			old_nrats = ratlist_length(&PCB->Data->Rat);
-			sprintf(buf, "%d rat%s", ratlist_length(&PCB->Data->Rat), ratlist_length(&PCB->Data->Rat) == 1 ? "" : "s");
+			sprintf(buf, "%ld rat%s", (long)ratlist_length(&PCB->Data->Rat), ratlist_length(&PCB->Data->Rat) == 1 ? "" : "s");
 			if (ratlist_length(&PCB->Data->Rat)) {
 				XtManageChild(XtParent(m_rats));
 				XtManageChild(m_rats);
